resource, utils refactor
