+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/_utils.py
@@ -17,13 +17,15 @@ from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.util import ns_to_iso_str
 from opentelemetry.util.types import Attributes
 
+from azure.core.pipeline.policies import BearerTokenCredentialPolicy
 from azure.monitor.opentelemetry.exporter._generated.models import ContextTagKeys, TelemetryItem
 from azure.monitor.opentelemetry.exporter._version import VERSION as ext_version
 from azure.monitor.opentelemetry.exporter._constants import (
-    _INSTRUMENTATIONS_BIT_MAP,
-    _WEBSITE_SITE_NAME,
-    _FUNCTIONS_WORKER_RUNTIME,
     _AKS_ARM_NAMESPACE_ID,
+    _APPLICATION_INSIGHTS_RESOURCE_SCOPE,
+    _INSTRUMENTATIONS_BIT_MAP,
+    _FUNCTIONS_WORKER_RUNTIME,
+    _WEBSITE_SITE_NAME,
 )
 
 
@@ -258,6 +260,19 @@ def _filter_custom_properties(properties: Attributes, filter=None) -> Dict[str, 
             continue
         truncated_properties[key] = str(val)[:8192]
     return truncated_properties
+
+
+def _get_auth_policy(credential, default_auth_policy):
+    if credential:
+        if hasattr(credential, 'get_token'):
+            return BearerTokenCredentialPolicy(
+                credential,
+                _APPLICATION_INSIGHTS_RESOURCE_SCOPE,
+            )
+        raise ValueError(
+            'Must pass in valid TokenCredential.'
+        )
+    return default_auth_policy
 
 
 class Singleton(type):
